TextEncodeBase

Documentation for TextEncodeBase.

TextEncodeBase.AbstractTokenizer

TextEncodeBase.splitting

TextEncodeBase.splitting

TextEncodeBase.tokenize

TextEncodeBase.tokenize

TextEncodeBase.AbstractTokenizer — Type

abstract type for tokenizers.

The tokenization procedure is separate into multiple TokenStages and recursive calls of splitting and tokenize. splitting break string into substrings, and tokenize is responsible for marking the TokenStages and do the recursive call.

TextEncodeBase.splitting — Function

splitting(tkr::AbstractTokenizer, x::TokenStages)

Split x given its tokenization stage. For example, the default behavior of a document stage is splitting into sentences (with WordTokenizers.split_sentences).

Overload this method for custom tokenizer.

TextEncodeBase.splitting — Method

splitting(tkr::AbstractTokenizer, s::TokenStages, x)

Interface for providing callback for splitting. x is the result of splitting(tkr, s).

Overload this method for custom splitting callback.

TextEncodeBase.tokenize — Function

tokenize(tkr::AbstractTokenizer, s::TokenStages, x)

Tokenize x. s contain the parent of x.

Overload this method when you need parent information to tokenize x.

TextEncodeBase.tokenize — Method

tokenize(tkr::AbstractTokenizer, x::TokenStages)

Tokenize x.

Overload this method for custom tokenizer and stages.