TextEncodeBase

Documentation for TextEncodeBase.

TextEncodeBase.AbstractTokenization

TextEncodeBase.AbstractTokenizer

TextEncodeBase.FlatTokenizer

TextEncodeBase.NestedTokenizer

TextEncodeBase.Splittability

TextEncodeBase.TokenStages

TextEncodeBase.Vocab

TextEncodeBase.lookup

TextEncodeBase.lookup

TextEncodeBase.lookup

TextEncodeBase.lookup

TextEncodeBase.nested2batch

TextEncodeBase.preprocess

TextEncodeBase.splittability

TextEncodeBase.splittable

TextEncodeBase.splitting

TextEncodeBase.tokenization

TextEncodeBase.tokenize_procedure

TextEncodeBase.trunc_and_pad

TextEncodeBase.with_head_tail

TextEncodeBase.wrap

TextEncodeBase.wrap

TextEncodeBase.AbstractTokenization — Type

abstract type for tokenization.

The tokenization procedure is separate into multiple TokenStages and recursive calls of splitting, wrap, and tokenize. splitting break string into substrings, wrap mark the substrings with new TokenStages, and tokenize is responsible for the tokenization.

TextEncodeBase.AbstractTokenizer — Type

abstract type for tokenizers.

Each tokenizer is link with a tokenization (by defining tokenization(::Tokenizer) = Tokenization()). The overall framework dispatch on both tokenizer and tokenization, but most of the time we only add methods for tokenization. This allow further composability and can interfere the tokenization process with given tokenizer.

TextEncodeBase.FlatTokenizer — Type

tokenizer that return flat array instead of nested array of tokens

TextEncodeBase.NestedTokenizer — Type

tokenizer that return nested array instead of flat array of tokens

TextEncodeBase.Splittability — Type

splittability trait

The splittability trait decide whether the given combination (tokenizer x tokenization x stage) is splittable or not (Splittable or UnSplittable). For example, DefaultTokenization and SentenceStage is splittable (i.e. splittability(::DefaultTokenization, ::SentenceStage) = Splittable()). The splittability change the behavior of tokenize: if it's splittable, tokenize will try to call splitting on the input, wrap each splitting result and recurse. Otherwise, it will directly call wrap and then recurse into tokenize.

TextEncodeBase.TokenStages — Type

abstract type for type that wrap input into specific stage for control tokenization.

There are six builtin stages in TextEncodeBase (all abstract XStage <: TokenStages):

1. Document <: DocumentStage: the input string is a full document,
 and thus need to be splitted into multiple sentence.
2. Sentence <: SentenceStage: the input string is a full string,
 and thus need to be splitted into multiple part (SubSentence/Word/Token).
3. SubSentence <: SubSentenceStage: special wrapper for case where the tokenizer
 does not directly break sentence all into words/tokens and these pieces contain
 multiple words/tokens, but you need the information that they are not full sentence.
4. Word <: WordStage: the input string is a single word.
5. SubWord <: SubWordStage: similar to SubSentence, but for word.
6. Token <: TokenStage: the final piece of the tokenization process.
 Generally, it's used to specify the end of this piece and should
 never be splitted.

Each wrapper have two field: x for the input, meta for extra information (nothing if not provided).

TextEncodeBase.Vocab — Type

Vocab(data::Vector{<:AbstractString}, unk::AbstractString="[UNK]")

Constructor for Vocab. data is the list of vocabulary word, can be nonunique. The actual list will be the unique version of data (i.e. vocab.list = unique(data)). unk is the indicator word for all unknown words. unk can be either in or not in data, depends on the use case.

TextEncodeBase.lookup — Function

lookup(v::Vocab, x)

Lookup x in v. lookup words depends on the type of x. If x is an integer, return the x-th word on the vocabulary list (i.e. v.list[x]) and return the unknown word if x is out-of-bound (v.unk). If x is a string, return the indice of x in the vocabulary list (i.e findfirst(==(x), v.list) and return the unknown indice if x not found in the list. If the unknown word v.unk is in the list, the unknown indice is its indice, otherwise 0.

Example

julia> vocab = Vocab(["a", "b", "c", "a", "b", "c"])
Vocab{String, StaticArrays.SizedVector{3, String, Vector{String}}}(size = 3, unk = [UNK], unki = 0)

julia> vocab_unk = Vocab(["a", "b", "xxx"], "xxx")
Vocab{String, StaticArrays.SizedVector{3, String, Vector{String}}}(size = 3, unk = xxx, unki = 3)

julia> lookup(vocab, "b")
2

julia> lookup(vocab, "d")
0

julia> lookup(vocab_unk, "d")
3

julia> lookup(vocab, 1)
"a"

julia> lookup(vocab, 10000)
"[UNK]"

julia> lookup(vocab_unk, 10000)
"xxx"

TextEncodeBase.lookup — Method

lookup(OneHot, v::Vocab, i)

lookup i and convert into one-hot representation.

Example

julia> lookup(OneHot, vocab, "a")
3-element OneHot{3}:
 1
 0
 0

julia> lookup(OneHot, vocab, ["a" "b"; "c" "d"])
3x2x2 OneHotArray{3, 3, Matrix{OneHot{0x00000003}}}:
[:, :, 1] =
 1  0
 0  0
 0  1

[:, :, 2] =
 0  0
 1  0
 0  0

julia> lookup(OneHot, vocab, 3)
ERROR: DomainError with c:
cannot convert `lookup(::Vocab, 3)` = "c" into one-hot representation.
Stacktrace:
[...]

TextEncodeBase.lookup — Method

lookup(v::Vocab, is::AbstractArray)

recursively lookup value from is

Example

julia> lookup(vocab, ["b", "c", "a", "A", "[UNK]"])
5-element Vector{Int64}:
 2
 3
 1
 0
 0

julia> lookup(vocab, [1, "a", 0, "A", "[UNK]"])
5-element Vector{Any}:
  "a"
 1
  "[UNK]"
 0
 0

TextEncodeBase.lookup — Method

lookup(v::Vocab, i::OneHotArray)

convert the one-hot representation back into words.

Example

julia> lookup(OneHot, vocab, ["a" "b"; "c" "d"])
3x2x2 OneHotArray{3, 3, Matrix{OneHot{0x00000003}}}:
[:, :, 1] =
 1  0
 0  0
 0  1

[:, :, 2] =
 0  0
 1  0
 0  0

julia> lookup(vocab, ans)
2×2 Matrix{String}:
 "a"  "b"
 "c"  "[UNK]"

TextEncodeBase.nested2batch — Method

nested2batch(x)

convert nested array into single array

Example

julia> TextEncodeBase.nested2batch([[[1 2],[3 4]]])
1×2×2×1 Array{Int64, 4}:
[:, :, 1, 1] =
 1  2

[:, :, 2, 1] =
 3  4

TextEncodeBase.preprocess — Method

preprocess(tkr::AbstractTokenizer, x)

Preprocess the input x. This is only called during tkr(x).

TextEncodeBase.splittability — Function

splittability(args...)

Return the splittability (Splittable/UnSplittable) of given argument combination. Overload to make a TokenStages splittable.

TextEncodeBase.splittable — Method

splittable(args...)

Return true if the splittability of given argument combination is Splittable().

TextEncodeBase.splitting — Function

splitting(t::AbstractTokenization, x::TokenStages)

Split x given its tokenization stage. For example, the default behavior of a document stage is splitting into sentences (with WordTokenizers.split_sentences).

Overload this method for custom tokenization.

TextEncodeBase.tokenization — Method

tokenization(::AbstractTokenizer) :: AbstractTokenization

Return the tokenization type of given tokenizer.

TextEncodeBase.tokenize_procedure — Method

tokenization_procedure(tokenizer, tokenizaton, stage)

The procedure of tokenization (splitting + wrap + tokenize).

TextEncodeBase.trunc_and_pad — Method

trunc_and_pad(x, n, pad)

truncate x to length n, otherwise add pad at the end of x until length equal n. x can be either nested or single array (but the element type should not be subtype of abstract array). if n is nothing, the largest length of the nested array will be used.

Example

julia> TextEncodeBase.trunc_and_pad(1:5, 7, -1)
7-element Vector{Int64}:
  1
  2
  3
  4
  5
 -1
 -1

julia> TextEncodeBase.trunc_and_pad([1:5, 2:7], 7, -1)
2-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, -1, -1]
 [2, 3, 4, 5, 6, 7, -1]

julia> TextEncodeBase.trunc_and_pad([1:5, [2:7, [1:2]]], nothing, -1)
2-element Vector{Vector}:
 [1, 2, 3, 4, 5, -1]
 Vector[[2, 3, 4, 5, 6, 7], [[1, 2, -1, -1, -1, -1]]]

TextEncodeBase.with_head_tail — Method

with_head_tail(x, head, tail)

Return [head; x; tail]. Ignored if head or tail is nothing.

Example

julia> TextEncodeBase.with_head_tail(1:5, -1, -2)
7-element Vector{Int64}:
 -1
  1
  2
  3
  4
  5
 -2

julia> TextEncodeBase.with_head_tail([1:5, 2:3], -1, -2)
2-element Vector{Vector{Int64}}:
 [-1, 1, 2, 3, 4, 5, -2]
 [-1, 2, 3, -2]

TextEncodeBase.wrap — Function

wrap(t::AbstractTokenization, s::TokenStages, x)

Mark the tokenization stage of x, which is part of the splitting result of s. For example, if we are doing simple whitespace tokenization and at the sentence stage, then x is just single word of s and thus return Word(x) (or Token(x)). Skip if x is already a TokenStages. (this method only apply to splittable stages)

Overload this method to control the tokenization process.

TextEncodeBase.wrap — Method

wrap(t::AbstractTokenization, x::TokenStages)

A handler for unsplittable stages (token/word/...).

Overload this method for custom transform.