TextEncodeBase

Documentation for TextEncodeBase.

TextEncodeBase.AbstractTokenization

TextEncodeBase.AbstractTokenizer

TextEncodeBase.NaiveIndexedMatchTokenizer

TextEncodeBase.NaiveIndexedTokenizer

TextEncodeBase.NaiveMatchTokenizer

TextEncodeBase.NaiveTokenizer

TextEncodeBase.TokenStages

TextEncodeBase.splitting

TextEncodeBase.splitting

TextEncodeBase.tokenize

TextEncodeBase.tokenize

TextEncodeBase.tokenize

TextEncodeBase.tokenize_procedure

TextEncodeBase.AbstractTokenization — Type

abstract type for tokenization.

The tokenization procedure is separate into multiple TokenStages and recursive calls of splitting and tokenize. splitting break string into substrings, and tokenize is responsible for marking each substring with a TokenStages and do the tokenization.

TextEncodeBase.AbstractTokenizer — Type

abstract type for tokenizers.

Each tokenizer is link with a tokenization (by defining tokenization(::Tokenizer) = Tokenization()). The overall framework dispatch on both tokenizer and tokenization, but most of the time we only add methods for tokenization. This allow further composability and can interfere the tokenization process with given tokenizer.

TextEncodeBase.NaiveIndexedMatchTokenizer — Type

default behavior but counting index and don't split some pattern

TextEncodeBase.NaiveIndexedTokenizer — Type

default behavior but counting the index

TextEncodeBase.NaiveMatchTokenizer — Type

default behavior but don't split some pattern

TextEncodeBase.NaiveTokenizer — Type

tokenizer that run the default behavior

TextEncodeBase.TokenStages — Type

abstract type for type that wrap input into specific stage for control tokenization.

There are six builtin stages in TextEncodeBase (all abstract XStage <: TokenStages):

1. Document <: DocumentStage: the input string is a full document,
 and thus need to be splitted into multiple sentence.
2. Sentence <: SentenceStage: the input string is a full string,
 and thus need to be splitted into multiple part (SubSentence/Word/Token).
3. SubSentence <: SubSentenceStage: special wrapper for case where the tokenizer
 does not directly break sentence all into words/tokens and these pieces contain
 multiple words/tokens, but you need the information that they are not full sentence.
4. Word <: WordStage: the input string is a single word.
5. SubWord <: SubWordStage: similar to SubSentence, but for word.
6. Token <: TokenStage: the final piece of the tokenization process.
 Generally, it's used to specify the end of this piece and should
 never be splitted.

Each wrapper have two field: x for the input, meta for extra information (nothing if not provided).

TextEncodeBase.splitting — Function

splitting(t::AbstractTokenization, x::TokenStages)

Split x given its tokenization stage. For example, the default behavior of a document stage is splitting into sentences (with WordTokenizers.split_sentences).

Overload this method for custom tokenization.

TextEncodeBase.splitting — Method

splitting(t::AbstractTokenization, s::TokenStages, x)

Interface for providing callback for splitting. x is the result of splitting(t, s).

Overload this method for custom splitting callback.

TextEncodeBase.tokenize — Function

tokenize(t::AbstractTokenization, s::TokenStages, x)

Mark the tokenization stage of x, which is part of the splitting result of s. For example, if we are doing simple whitespace tokenization and at the sentence stage, then x is just single word of s and thus return Word(x) (or Token(x)). Skip if x is already a TokenStages. (this method only apply to splittable stages)

Overload this method to control the tokenization process.

TextEncodeBase.tokenize — Method

tokenize(t::AbstractTokenization, x::TokenStages)

A handler for unsplittable stages (token/word/...).

Overload this method for custom transform.

TextEncodeBase.tokenize — Method

tokenize(tkr::AbstractTokenizer, t::AbstractTokenization, x::TokenStages)

Tokenize x according to tkr and t.

Overload for custom tokenizer, tokenization and stages. For making a unsplittable into splittable (or vice versa), you must overload this method.

TextEncodeBase.tokenize_procedure — Method

tokenization_procedure(tokenizer, tokenizaton, stage)

The procedure of tokenization (splitting + tokenize). This is use to restore full behavior for stage that default unsplittable. Generally don't overload this function.